addnig answers repaired

+++ src/main/java/pl/micsoc/dictionary/repository/AnswerRepository.java
@@ -3,6 +3,6 @@ package pl.micsoc.dictionary.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.micsoc.dictionary.model.Answer;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
 
 }
